+++ frontend/package.json
@@ -22,6 +22,7 @@
     "eslint-plugin-react": "^7.33.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "prettier": "^3.0.0",
+    "sass": "^1.69.7",
     "vite": "^4.4.7"
   },
   "lint-staged": {
