+++ frontend/src/components/MessageParser.jsx
