+++ frontend/src/components/choix/productChoise.scss
@@ -1,0 +1,13 @@
+.learning-options-container {
+  display: flex;
+  align-items: flex-start;
+  flex-wrap: wrap;
+}
+
+.learning-option-button {
+  padding: 0.246rem;
+  border-radius: 25px;
+  background: transparent;
+  border: 1px solid green;
+  margin: 0.2rem;
+}
